AURORA_APP/requirements.txt
edge-tts>=6.1.0
python-telegram-bot[job-queue]>=21.0
python-dotenv>=1.0.0
streamlit>=1.28.0